Refactor: Phase 2 - Tool Modularization
- Created tools directory structure
- Migrated and refactored tools into separate modules
- Updated server.py to register tool modules
