+++ src/canvas_mcp/tools/__init__.py
@@ -1,0 +1,18 @@
+"""
+Canvas MCP Tools
+
+This package contains the tool functions for the Canvas MCP server.
+Each module in this package contains related tool functions that are
+registered with the MCP server.
+"""
+
+__all__ = [
+    "assignments",
+    "announcements",
+    "courses",
+    "files",
+    "modules",
+    "search",
+    "syllabus",
+    "sync",
+]
